Revert move of singleHackerView
